Remove pseudo-interface for master.py + small fixes + cleanup
