+++ bci/web/blueprints/api.py
@@ -3,11 +3,15 @@ import logging
 import os
 import threading
 
-from flask import Blueprint, request
+from flask import Blueprint, current_app, request
 
+import bci.browser.support as browser_support
+import bci.evaluations.logic as application_logic
+from bci.analysis.plot_factory import PlotFactory
 from bci.app import sock
-from bci.evaluations.logic import evaluation_factory
-from bci.main import Main as bci_api
+from bci.configuration import Global, Loggers
+from bci.evaluations.logic import PlotParameters
+from bci.main import Main
 from bci.web.clients import Clients
 
 logger = logging.getLogger(__name__)
@@ -28,14 +32,24 @@ def __start_thread(func, args=None) -> bool:
         return True
 
 
+def __get_main() -> Main:
+    if main := current_app.config['main']:
+        return main
+    raise Exception('Main object is not instantiated')
+
+
 @api.before_request
 def check_readiness():
-    if not bci_api.is_ready():
+    try:
+        pass
+        # _ = ____get_main()
+    except Exception as e:
+        logger.critical(e)
         return {
             'status': 'NOK',
             'msg': 'BugHog is not ready',
             'info': {
-                'log': bci_api.get_logs(),
+                'log': Loggers.get_logs()
             }
         }
 
@@ -63,8 +77,8 @@ def start_evaluation():
         }
 
     data = request.json.copy()
-    params = evaluation_factory(data)
-    if __start_thread(bci_api.run, args=[params]):
+    params = application_logic.evaluation_factory(data)
+    if __start_thread(__get_main().run, args=[params]):
         return {
             'status': 'OK'
         }
@@ -85,9 +99,9 @@ def stop_evaluation():
     data = request.json.copy()
     forcefully = data.get('forcefully', False)
     if forcefully:
-        bci_api.stop_forcefully()
+        __get_main().activate_stop_forcefully()
     else:
-        bci_api.stop_gracefully()
+        __get_main().activate_stop_gracefully()
     return {
         'status': 'OK'
     }
@@ -113,7 +127,7 @@ def init_websocket(ws):
             if params := message.get('select_project', None):
                 Clients.associate_project(ws, params)
             if requested_variables := message.get('get', []):
-                Clients.push_info(ws, *requested_variables)
+                __get_main().push_info(ws, *requested_variables)
         except ValueError:
             logger.warning('Ignoring invalid message from client.')
     ws.send('Connected to BugHog')
@@ -123,7 +137,7 @@ def init_websocket(ws):
 def get_browsers():
     return {
         'status': 'OK',
-        'browsers': bci_api.get_browser_support()
+        'browsers': [browser_support.get_chromium_support(), browser_support.get_firefox_support()]
     }
 
 
@@ -131,7 +145,7 @@ def get_browsers():
 def get_projects():
     return {
         'status': 'OK',
-        'projects': bci_api.get_projects_of_custom_framework()
+        'projects': __get_main().evaluation_framework.get_projects()
     }
 
 
@@ -145,7 +159,7 @@ def create_project():
     project_name = request.json.get('project_name')
     return {
         'status': 'OK',
-        'projects': bci_api.create_empty_project(project_name)
+        'projects': __get_main().evaluation_framework.create_empty_project(project_name)
     }
 
 
@@ -174,23 +188,21 @@ def data_source():
         }
 
     params = request.json.copy()
-    revision_data, version_data = bci_api.get_data_sources(params)
-    if revision_data or version_data:
-        return {
-            'status': 'OK',
-            'revision': revision_data,
-            'version': version_data
-        }
-    else:
+    plot_params = PlotParameters.from_dict(params)
+    if missing_params := PlotFactory.validate_params(plot_params):
         return {
             'status': 'NOK',
-            'msg': 'Invalid type'
+            'msg': f'Missing plot parameters: {missing_params}'
         }
-
+    return {
+            'status': 'OK',
+            'revision': PlotFactory.get_plot_revision_data(params),
+            'version': PlotFactory.get_plot_version_data(params)
+        }
 
 @api.route('/poc/<string:project>/', methods=['GET'])
 def get_experiments(project: str):
-    experiments = bci_api.get_mech_groups_of_evaluation_framework('custom', project)
+    experiments = __get_main().evaluation_framework.get_mech_groups(project)
     return {
         'status': 'OK',
         'experiments': experiments
@@ -201,7 +213,7 @@ def get_experiments(project: str):
 def poc(project: str, poc: str):
     return {
         'status': 'OK',
-        'tree': bci_api.get_poc(project, poc)
+        'tree': __get_main().evaluation_framework.get_poc_structure(project, poc)
     }
 
 
@@ -212,7 +224,7 @@ def get_poc_file_content(project: str, poc: str, file: str):
     if request.method == 'GET':
         return {
             'status': 'OK',
-            'content': bci_api.get_poc_file(project, poc, domain, path, file)
+            'content': __get_main().evaluation_framework.get_poc_file(project, poc, domain, path, file)
         }
     else:
         if not request.json:
@@ -221,7 +233,8 @@ def get_poc_file_content(project: str, poc: str, file: str):
                 'msg': 'No content to update file with'
             }
         data = request.json.copy()
-        success = bci_api.update_poc_file(project, poc, domain, path, file, data['content'])
+        content = data['content']
+        success = __get_main().evaluation_framework.update_poc_file(project, poc, domain, path, file, content)
         if success:
             return {
                 'status': 'OK'
@@ -241,7 +254,10 @@ def add_page(project: str, poc: str):
         }
 
     data = request.json.copy()
-    success = bci_api.add_page(project, poc, data['domain'], data['page'], data['file_type'])
+    domain = data['domain']
+    path = data['page']
+    file_type = data['file_type']
+    success = __get_main().evaluation_framework.add_page(project, poc, domain, path, file_type)
     if success:
         return {
             'status': 'OK'
@@ -260,7 +276,8 @@ def add_config(project: str, poc: str):
             'msg': "No parameters found"
         }
     data = request.json.copy()
-    success = bci_api.add_config(project, poc, data['type'])
+    type = data['type']
+    success = __get_main().evaluation_framework.add_config(project, poc, type)
     if success:
         return {
             'status': 'OK'
@@ -275,7 +292,7 @@ def add_config(project: str, poc: str):
 def get_available_domains():
     return {
         'status': 'OK',
-        'domains': bci_api.get_available_domains()
+        'domains': Global.get_available_domains()
     }
 
 
@@ -293,7 +310,8 @@ def create_experiment(project: str):
             'status': 'NOK',
             'msg': 'Missing experiment name'
         }
-    if bci_api.create_empty_poc(project, data['poc_name']):
+    poc_name = data['poc_name']
+    if __get_main().evaluation_framework.create_empty_poc(project, poc_name):
         return {
             'status': 'OK'
         }
