+++ bci/web/vue/src/components/poc-editor.vue
@@ -286,7 +286,6 @@
         this.active_poc.name = null;
         this.active_poc.active_domain = null;
         this.active_poc.active_path = null;
-        this.update_poc_tree(val);
       },
     },
   };
